scripts/contract_parser.py
# ...existing code...
import os
import fitz
import docx


def extract_text_from_pdf(pdf_path):
    doc = fitz.open(pdf_path)
    text = ""
    for page in doc:
        text += page.get_text()
    return text.strip()

def extract_text_from_docx(docx_path):
    doc = docx.Document(docx_path)
    return "\n".join([para.text for para in doc.paragraphs]).strip()

def extract_text_from_file(file_path):
    with open(file_path, "r", encoding="utf-8") as file:
        return file.read().strip()

def extract_contract_text(file_path):
    ext  = os.path.splitext(file_path)[1].lower()

    if ext == ".pdf":
        return extract_text_from_pdf(file_path)
    elif ext == ".docx":
        return extract_text_from_docx(file_path)
    elif ext in [".txt"]:
        return extract_text_from_file(file_path)
    else:
        raise ValueError(f"Unsupported file type: {ext}. Supported types are: .pdf, .docx, .txt")
# ...existing code...
