requirements.txt
python-docx
PyMuPDF
requests
